chore: move entry-rsc
